# [CN] improve CN performance