--- a/backend/cn/lib/typing.ml
+++ b/backend/cn/lib/typing.ml
@@ -445,7 +445,11 @@ let add_r_internal loc (r, RE.O oargs) =
   let@ simp_ctxt = simp_ctxt () in
   let r = Simplify.ResourceTypes.simp simp_ctxt r in
   let oargs = Simplify.IndexTerms.simp simp_ctxt oargs in
-  let pointer_facts = Resources.pointer_facts ((r, RE.O oargs) :: Context.get_rs s) in
+  let pointer_facts =
+    Resources.pointer_facts
+      ~new_resource:(r, RE.O oargs)
+      ~old_resources:(Context.get_rs s)
+  in
   let@ () = set_typing_context (Context.add_r loc (r, O oargs) s) in
   iterM (fun x -> add_c_internal (LC.T x)) pointer_facts