Performance improvement for CN, by reducing the duplication in the resource-derived constraints. 

Before
```
11:23 % time cn verify --only own_backwards_append_lemma src/examples/Dbl_Queue/lemmas_other.c
[1/1]: own_backwards_append_lemma
cn verify --only own_backwards_append_lemma   36.20s user 12.47s system 79% cpu 1:01.61 total
```

After
```
11:24 % time cn verify --only own_backwards_append_lemma src/examples/Dbl_Queue/lemmas_other.c
[1/1]: own_backwards_append_lemma
cn verify --only own_backwards_append_lemma   9.61s user 4.63s system 54% cpu 26.070 total